Add player superclass for Human and AI to inherit from
